Fold the getAs/getManyAs methods into get/getMany
Using the transform and then operation afterwards rather than changing
in place means that different calls to the many methods can provide
different types (for the same stream). It also makes the code simpler,
and only has the overhead of retaining the js.Proxy data for an
additional method call.