BREAKING CHANGE Compilation of polymer
The compiled to javascript version does not work on firefox. Complains
about:

SyntaxError: missing ) after formal parameters @example.html_bootstrap.dart.js:145835

This is in code which works just fine in Dart.

Furthermore, this relocates the example.html link. That should be fixed
to the original location to make sure that broken links are not
produced.