Twitter are such killjoys
The api is now rate limited and requires a user or app token. Since I
don't really want to sign up to twitter just to enable this app it means
the example is now broken.

Changing this URL at least shows the 400 rather than a misleading 404.

Going to alter the app to spell out this new behaviour.